Implemented authorization
- admin can access everything
- users can add and may only edit or delete their own posts

+++ src/Controller/MediaController.php
@@ -2,17 +2,14 @@
 
 namespace App\Controller;
 
-use App\Utility\MediaHelper;
-
-class MediaController extends MediaHelper {
+class MediaController extends AppController {
 
     public function initialize() {
         parent::initialize();
         $this->loadModel( 'Media' );
-        // temporary until authentication is done
+        // temporary until authorization is done
         $this->loadModel( 'Users' );
-        // temporary until checks for file type is implemented
-        $this->loadModel( 'MediaTypes' );
+        $this->loadComponent( 'MediaHelper' );
     }
 
     public function delete( $id ) {
@@ -23,11 +20,17 @@ class MediaController extends MediaHelper {
         if ( $this->Media->delete( $media ) ) {
             // remove files from filesystem
             unlink( WWW_ROOT . 'img/' . $media['media_link'] );
-            unlink( WWW_ROOT . 'img/' . $media['thumb_link'] );
-            $this->Flash->success( __( 'The article with id: {0} has been deleted.',
-                            h( $id ) ) );
-            return $this->redirect( ['controller' => 'Results', 'action' => 'search'] );
+            if ( $media['thumb_link'] != null ) {
+                unlink( WWW_ROOT . 'img/' . $media['thumb_link'] );
+            }
+            return $this->redirect( ['action' => 'posts'] );
         }
+    }
+    
+    public function posts () {
+        $userProducts = $this->Media->find('all')
+                ->where(['author_id' => $this->Auth->user('user_id')]);
+        $this->set(compact('userProducts'));
     }
 
     public function edit( $id ) {
@@ -40,58 +43,72 @@ class MediaController extends MediaHelper {
         if ( $this->request->is( 'post' ) ) {
 
             $input = $this->request->data;
-            $searchGenreId = $input['genre_id'];
-            $searchGenreName = strtolower( $this->getGenreName( $searchGenreId ) );
+            $input['author_id'] = $this->Auth->user( 'user_id' );
 
-            $mediaName = strtolower($input['file']['name']);
-            
+            $searchGenreId = $input['genre_id'];
+            $searchGenreName = strtolower( $this->MediaHelper->getGenreName( $searchGenreId ) );
+
+            $mediaName = strtolower( $input['file']['name'] );
+
             // add "unique" string to the name of the file to avoid over writes
             $mediaStoredName = uniqid() . '-' . $mediaName;
 
             // path link for full image that will be stored in the database
             $mediaPathLink = 'media/' . $searchGenreName . '/' . $mediaStoredName;
 
-            // path link for thumbnail that will be stored in the database
-            $mediaThumbLink = 'media/' . $searchGenreName . '/' . 'thumbnail-' . $mediaStoredName;
-
             // the absolute path where the media will be stored
             $storedPath = WWW_ROOT . 'img/' . $mediaPathLink;
 
             $input['media_link'] = $mediaPathLink;
-            $input['thumb_link'] = $mediaThumbLink;
 
-            $newMedia = $this->Media->patchEntity( $newMedia, $input );
+            $finfo = finfo_open();
+            $mime = strtolower( finfo_file( $finfo, $input['file']['tmp_name'], FILEINFO_MIME ) );
+            finfo_close( $finfo );
 
-            if ( $this->Media->save( $newMedia ) ) {
-                // generate and store thumbnail it. Also store uploaded file
-                $this->generateThumbnail( $input['file']['tmp_name'],
-                        $mediaThumbLink );
-                move_uploaded_file( $input['file']['tmp_name'], $storedPath );
+            if ( strstr( $mime, 'image/' ) ) {
+                // path link for thumbnail that will be stored in the database
+                $mediaThumbLink = 'media/' . $searchGenreName . '/' . 'thumbnail-' . $mediaStoredName;
+                $input['type_id'] = 1;
+                $input['thumb_link'] = $mediaThumbLink;
+
+                $newMedia = $this->Media->patchEntity( $newMedia, $input );
+
+                if ( $this->Media->save( $newMedia ) ) {
+                    // generate and store thumbnail it. Also store uploaded file
+                    $this->generateThumbnail( $input['file']['tmp_name'], $mediaThumbLink );
+                    move_uploaded_file( $input['file']['tmp_name'], $storedPath );
+                } else {
+                    $this->Flash->error( __( 'The media could not be saved. Please, try again.' ) );
+                }
+                // redirect to view later
+                return $this->redirect( ['action' => 'posts'] );
             } else {
-                $this->Flash->error( __( 'The media could not be saved. Please, try again.' ) );
+                $input['type_id'] = 2;
+                $newMedia = $this->Media->patchEntity( $newMedia, $input );
+
+                if ( $this->Media->save( $newMedia ) ) {
+                    move_uploaded_file( $input['file']['tmp_name'], $storedPath );
+                } else {
+                    $this->Flash->error( __( 'The media could not be saved. Please, try again.' ) );
+                }
+                // redirect to view later
+                return $this->redirect( ['action' => 'posts'] );
             }
-            return $this->redirect( '' );
         }
 
-        $genreList = $this->getGenreList();
+        $genreList = $this->MediaHelper->getGenreList();
 
-        // temporary until authentication can grab user's id
-        $userList = $this->Users->find( 'list',
-                ['keyField' => 'user_id', 'valueField' => 'username'] );
 
-        // temporary until file can be checked for video or image
-        $typeList = $this->MediaTypes->find( 'list',
-                ['keyField' => 'type_id', 'valueField' => 'type_name'] );
-
-        $this->set( compact( 'genreList', 'newMedia', 'userList', 'typeList' ) );
+        $this->set( compact( 'genreList', 'newMedia' ) );
     }
 
-    private function generateThumbnail( $source, $mediaThumbLink ) {
+    private function generateThumbnail( $source, $mediaThumbLink) {
         $nw = 300;
         $nh = 300;
         $imageInfo = getimagesize( $source );
         $w = $imageInfo[0];
         $h = $imageInfo[1];
+
         $type = strtolower($imageInfo['mime']);
 
         switch ( $type ) {
@@ -122,14 +139,12 @@ class MediaController extends MediaHelper {
             $half_width = $adjusted_width / 2;
             $int_width = $half_width - $w_height;
 
-            imagecopyresampled( $dimg, $simg, -$int_width, 0, 0, 0,
-                    $adjusted_width, $nh, $w, $h );
+            imagecopyresampled( $dimg, $simg, -$int_width, 0, 0, 0, $adjusted_width, $nh, $w, $h );
         } elseif ( ($w < $h) || ($w == $h) ) {
             $adjusted_height = $h / $wm;
             $half_height = $adjusted_height / 2;
             $int_height = $half_height - $h_height;
-            imagecopyresampled( $dimg, $simg, 0, -$int_height, 0, 0, $nw,
-                    $adjusted_height, $w, $h );
+            imagecopyresampled( $dimg, $simg, 0, -$int_height, 0, 0, $nw, $adjusted_height, $w, $h );
         } else {
             imagecopyresampled( $dimg, $simg, 0, 0, 0, 0, $nw, $nh, $w, $h );
         }
@@ -152,6 +167,24 @@ class MediaController extends MediaHelper {
 
         imagedestroy( $simg );
         imagedestroy( $dimg );
+    }
+
+    public function isAuthorized( $user ) {
+        // All registered users can add articles
+
+        if ( in_array( $this->request->getParam( 'action' ), ['add', 'posts'] ) ) {
+            return true;
+        }
+        
+        // The owner of an article can edit and delete it
+        if ( in_array( $this->request->getParam( 'action' ), ['edit', 'delete'] ) ) {
+            $articleId = (int) $this->request->getParam( 'pass.0' );
+            if ( $this->Media->isOwnedBy( $articleId, $user['user_id'] ) ) {
+                return true;
+            }
+        }
+
+        return parent::isAuthorized( $user );
     }
 
 }
